Hadoop: Move HadoopExploitClient.exploit_urls() to HadoopExploiter
