+++ monkey/agent_plugins/exploiters/hadoop/src/hadoop_exploit_client.py
@@ -3,7 +3,7 @@ import logging
 import posixpath
 import string
 from time import time
-from typing import Sequence, Tuple
+from typing import Tuple
 from uuid import UUID
 
 import requests
@@ -17,7 +17,6 @@ from common.tags import (
 )
 from common.types import Event
 from common.utils.code_utils import insecure_generate_random_string
-from infection_monkey.i_puppet import ExploiterResultData
 from infection_monkey.model import TargetHost
 
 from .hadoop_options import HadoopOptions
@@ -47,27 +46,6 @@ class HadoopExploitClient:
     ):
         self._agent_id = agent_id
         self._agent_event_publisher = agent_event_publisher
-
-    def exploit_urls(
-        self,
-        target_host: TargetHost,
-        options: HadoopOptions,
-        interrupt: Event,
-        urls: Sequence[str],
-        command: str,
-    ):
-        """
-        Attempt to exploit on all URLs
-        """
-        exploit_result = ExploiterResultData()
-
-        for url in urls:
-            if self.exploit(target_host, options, interrupt, url, command):
-                exploit_result.exploitation_success = True
-                exploit_result.propagation_success = True
-                break
-
-        return exploit_result
 
     def exploit(
         self,
