+++ monkey/agent_plugins/exploiters/hadoop/src/hadoop_exploiter.py
@@ -75,6 +75,27 @@ class HadoopExploiter:
                 agent_binary_http_server, options.agent_binary_download_timeout
             )
 
+    def exploit_urls(
+        self,
+        target_host: TargetHost,
+        options: HadoopOptions,
+        interrupt: Event,
+        urls: Sequence[str],
+        command: str,
+    ) -> ExploiterResultData:
+        """
+        Attempt to exploit on all URLs
+        """
+        exploit_result = ExploiterResultData()
+
+        for url in urls:
+            if self._hadoop_exploit_client.exploit(target_host, options, interrupt, url, command):
+                exploit_result.exploitation_success = True
+                exploit_result.propagation_success = True
+                break
+
+        return exploit_result
+
 
 def _build_potential_urls(target_host: TargetHost, options: HadoopOptions) -> Sequence[str]:
     return build_urls(str(target_host.ip), [(str(p), False) for p in options.target_ports])
