+++ monkey/tests/unit_tests/agent_plugins/exploiters/hadoop/test_hadoop_exploit_client.py
@@ -218,31 +218,3 @@ def create_agent_binary_server(
     host, agent_binary_repository
 ) -> Tuple[Optional[str], Optional[LockedHTTPServer]]:
     return "url", MagicMock()
-
-
-def test_exploit_urls__success(mock_hadoop_server, exploit: Callable[[], bool]):
-    result = hadoop_exploiter_client.exploit_urls(URLS_TO_EXPLOIT, COMMAND)
-
-    assert result.exploitation_success
-    assert result.propagation_success
-
-
-def test_exploit_urls__fails_if_no_urls(mock_hadoop_server, exploit: Callable[[], bool]):
-    result = hadoop_exploiter_client.exploit_urls([], COMMAND)
-
-    assert not result.exploitation_success
-    assert not result.propagation_success
-
-
-def test_exploit_urls__attempts_each_url(mock_hadoop_server, exploit: Callable[[], bool]):
-    mock_hadoop_server.post(SUBMIT_APP_URL_1, json="", status_code=500)
-    mock_hadoop_server.post(SUBMIT_APP_URL_2, json="", status_code=500)
-
-    hadoop_exploiter_client.exploit_urls(URLS_TO_EXPLOIT, COMMAND)
-
-    # Check requests to see which URLs are attempted
-    called_urls = [h.url for h in mock_hadoop_server.request_history]
-    assert NEW_APP_URL_1 in called_urls
-    assert SUBMIT_APP_URL_1 in called_urls
-    assert NEW_APP_URL_2 in called_urls
-    assert SUBMIT_APP_URL_2 in called_urls
